+++ modules/token/issue_token.go
@@ -9,8 +9,8 @@ type DocMsgIssueToken struct {
 	Name          string `bson:"name"`
 	Scale         uint32 `bson:"scale"`
 	MinUnit       string `bson:"min_unit"`
-	InitialSupply uint64 `bson:"initial_supply"`
-	MaxSupply     uint64 `bson:"max_supply"`
+	InitialSupply int64  `bson:"initial_supply"`
+	MaxSupply     int64  `bson:"max_supply"`
 	Mintable      bool   `bson:"mintable"`
 	Owner         string `bson:"owner"`
 }
@@ -26,9 +26,9 @@ func (m *DocMsgIssueToken) BuildMsg(v interface{}) {
 	m.Name = msg.Name
 	m.Scale = msg.Scale
 	m.MinUnit = msg.MinUnit
-	m.InitialSupply = msg.InitialSupply
+	m.InitialSupply = int64(msg.InitialSupply)
 	m.Owner = msg.Owner
-	m.MaxSupply = msg.MaxSupply
+	m.MaxSupply = int64(msg.MaxSupply)
 	m.Mintable = msg.Mintable
 }
